+++ test/sql/adm1.sql
@@ -414,6 +414,7 @@ select emaj.emaj_force_drop_group('dummyGroup');
 select emaj.emaj_get_previous_mark_group('dummyGroup', '2010-01-01');
 select emaj.emaj_get_previous_mark_group('dummyGroup', 'EMAJ_LAST_MARK');
 select * from emaj.emaj_log_stat_group('dummyGroup', 'dummyMark', NULL); 
+select * from emaj.emaj_log_stat_groups(array['dummyGroup'], 'dummyMark', NULL); 
 select * from emaj.emaj_detailed_log_stat_group('dummyGroup', 'dummyMark', NULL);
 select emaj.emaj_estimate_rollback_group('dummyGroup', 'dummyMark', TRUE);
 select emaj.emaj_estimate_rollback_groups(array['dummyGroup'], 'dummyMark', FALSE);