+++ test/sql/misc.sql
@@ -69,11 +69,11 @@ select emaj.emaj_set_mark_group('myGroup2','Mark23');
 select emaj.emaj_set_mark_groups(array['myGroup1','myGroup2'],'Multi-3');
 
 -----------------------------
--- emaj_log_stat_group() and emaj_detailled_log_stat_group() test
+-- emaj_log_stat_group(), emaj_log_stat_groups(), emaj_detailled_log_stat_group() and emaj_detailled_log_stat_groups() test
 -----------------------------
 -- group is unknown in emaj_group_def
 select * from emaj.emaj_log_stat_group(NULL,NULL,NULL);
-select * from emaj.emaj_log_stat_group('unknownGroup',NULL,NULL);
+select * from emaj.emaj_log_stat_groups(array['unknownGroup'],NULL,NULL);
 select * from emaj.emaj_detailed_log_stat_group(NULL,NULL,NULL);
 select * from emaj.emaj_detailed_log_stat_group('unknownGroup',NULL,NULL);
 
@@ -82,6 +82,9 @@ select * from emaj.emaj_log_stat_group('myGroup2','dummyStartMark',NULL);
 select * from emaj.emaj_log_stat_group('myGroup2',NULL,'dummyEndMark');
 select * from emaj.emaj_detailed_log_stat_group('myGroup2','dummyStartMark',NULL);
 select * from emaj.emaj_detailed_log_stat_group('myGroup2',NULL,'dummyEndMark');
+
+select * from emaj.emaj_log_stat_groups(array['myGroup1','myGroup2'],NULL,NULL);
+select * from emaj.emaj_log_stat_groups(array['myGroup1','myGroup2'],'Multi-1','dummyEndMark');
 
 -- start mark > end mark
 -- original test (uncomment for unit test)
@@ -129,6 +132,13 @@ select stat_group, stat_schema, stat_table, stat_first_mark, stat_last_mark, sta
 select stat_group, stat_schema, stat_table, stat_first_mark, stat_last_mark, stat_rows from emaj.emaj_log_stat_group('myGroup2','EMAJ_LAST_MARK','')
   order by stat_group, stat_schema, stat_table;
 
+select stat_group, stat_schema, stat_table, stat_first_mark, stat_last_mark, stat_rows
+  from emaj.emaj_log_stat_groups(array['myGroup1','myGroup2'],'Multi-1',NULL)
+  order by stat_group, stat_schema, stat_table;
+select stat_group, stat_schema, stat_table, stat_first_mark, stat_last_mark, stat_rows
+  from emaj.emaj_log_stat_groups(array['myGroup1','myGroup2'],'Multi-1','Multi-3')
+  order by stat_group, stat_schema, stat_table;
+
 select stat_group, stat_schema, stat_table, stat_first_mark, stat_last_mark, stat_role, stat_verb, stat_rows
   from emaj.emaj_detailed_log_stat_group('myGroup2',NULL,NULL)
   order by stat_group, stat_schema, stat_table;
@@ -162,7 +172,7 @@ select * from emaj.emaj_detailed_log_stat_group('emptyGroup',NULL,NULL);
 begin;
   select emaj.emaj_stop_group('myGroup1');
   select emaj.emaj_reset_group('myGroup1');
-  select * from emaj.emaj_log_stat_group('myGroup1',NULL,NULL);
+  select * from emaj.emaj_log_stat_groups(array['myGroup1'],NULL,NULL);
   select * from emaj.emaj_detailed_log_stat_group('myGroup1',NULL,NULL);
 rollback;